pyproject.toml
[project]
name = "sg-jobs-agent"
version = "0.1.0"
description = "Daily digest of Singapore tech roles matched to a personal profile"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.10",
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
